API/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
xmltodict==0.12.0
boto3==1.26.137
botocore==1.29.137
python-dotenv==1.0.0
